+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Guardian Dashboard Release Notes
 
-## 0.0.2 (2021-11-??)
+## 0.0.3 (2021-12-??)
+
+* **bug:** Fixed the playing audio functionality in Safari ([CE-1553](https://jira.rfcx.org/browse/CE-1553))
+* **feature:** #25 Display the `immediate` label if the difference of calculation time to receive less than 60s ([#25](https://github.com/rfcx/incident-center/issues/25))
+* **bug:** Show email when firstname and lastname are not defined ([#28](https://github.com/rfcx/incident-center/issues/28))
+
+## 0.0.2 (2021-11-30)
 
 * **feature:** Pagination is added to the homepage ([CE-1569](https://jira.rfcx.org/browse/CE-1569))
 * **feature:** Resolution time is added to closed reports ([CE-1572](https://jira.rfcx.org/browse/CE-1572))
@@ -11,7 +17,6 @@
 
 ## 0.0.1 (2021-10-24)
 
-* **bug:** Fixed the playing audio functionality in Safari ([CE-1553](https://jira.rfcx.org/browse/CE-1553))
 * **bug:** Response assets are loaded only when user opens them ([CE-1515](https://jira.rfcx.org/browse/CE-1515))
 * **bug:** Fixed the datetime format on the incident page for event start/end time ([CE-1516](https://jira.rfcx.org/browse/CE-1516))
 * **feature:** The title is added to homepage for non-authenticated state ([CE-1501](https://jira.rfcx.org/browse/CE-1501))